# bpo-40255: Implement Immortal Instances - Optimizations Combined