… Lib/test/test_gc.py
… Lib/test/test_io.py
… Lib/test/test_regrtest.py
… Lib/test/test_sys.py
… Lib/test/test_threading.py
… Lib/test/test_warnings/__init__.py
… Python/pylifecycle.c