@kumaraditya303 2022-02-22
You can use this in deepfrozen modules to get this even faster see https://github.com/python/cpython/blob/74127b89a8224d021fc76f679422b76510844ff9/Tools/scripts/deepfreeze.py#L140