@ericsnowcurrently 2022-02-22
> Overall: 0% faster compared to the main branch

That's great news!  I'm going to update PEP 683 with the outcome and some of the details.